First step to break the monolith. Splitting into unit classes.
