+++ web/premises/managers/__init__.py
@@ -1,0 +1,2 @@
+from delete_prevention import DeletePreventionManager
+from contention import ContentionManager
