+++ web/premises/mixins/form_renderer.py
@@ -1,0 +1,9 @@
+class FormRenderer(object):
+    def as_p(self):
+        "Returns this form rendered as HTML <p>s."
+        return self._html_output(
+            normal_row='<p%(html_class_attr)s>%(label)s %(field)s%(help_text)s</p>',
+            error_row='%s',
+            row_ender='</p>',
+            help_text_html=' <div class="helptext">%s</div>',
+            errors_on_separate_row=True)
