+++ web/premises/models/comment.py
@@ -1,0 +1,16 @@
+from . import Premise
+
+from django.db import models
+from django.conf import settings
+from django.utils.encoding import smart_unicode
+
+
+class Comment(models.Model):
+    premise = models.ForeignKey(Premise)
+    user = models.ForeignKey(settings.AUTH_USER_MODEL)
+    text = models.TextField()
+    date_creation = models.DateTimeField(auto_now_add=True)
+    is_active = models.BooleanField(default=True)
+
+    def __unicode__(self):
+        return smart_unicode(self.text)
